API-prisma
- Added ts
