+++ src/index.ts
@@ -1,4 +1,4 @@
-const app = require('./server');
+import app from './server';
 
 app.listen(3005, () => {
   console.log("Server running on http://localhost:3005");
